Liquibase schema and data

+++ src/main/java/pl/dmichalski/reservations/business/domain/entity/room/RoomXReservationEntity.java
@@ -13,7 +13,7 @@ public class RoomXReservationEntity {
     @EmbeddedId
     private RoomXReservationPK roomXReservationPK;
 
-    private long roomPrice;
+    private Integer roomPrice;
 
     public RoomXReservationEntity() {
     }
@@ -42,7 +42,7 @@ public class RoomXReservationEntity {
         return roomPrice;
     }
 
-    public void setRoomPrice(long roomPrice) {
+    public void setRoomPrice(Integer roomPrice) {
         this.roomPrice = roomPrice;
     }
 }
